moved create_key / create_character_set and create_key_hash methods into shift class - tests created and passing
