+++ lib/shift.rb
@@ -1,0 +1,29 @@
+class Shift
+  def initialize
+  end
+
+###### create_character_set creates an alphabet array ######
+###### with a space character included ######
+
+  def create_character_set
+    ("a".."z").to_a << " "
+  end
+
+###### create_key generates a 0 padded 5 different random key string ######
+
+  def create_key
+  string_integer = (4.times.map{rand(10)}.join)
+  string_integer.to_s.rjust(5, "0")
+  end
+
+  def create_key_hash(key)
+  split_key = key.split("").map {|chr| chr.to_i}
+  letters = ["A", "B", "C", "D"]
+  key_hash = {}
+  key_hash["A"] = split_key[0..1]
+  key_hash["B"] = split_key[1..2]
+  key_hash["C"] = split_key[2..3]
+  key_hash["D"] = split_key[3..4]
+  key_hash
+  end
+end
