+++ test/enigma_test.rb
@@ -1,5 +1,6 @@
 require "./test/test_helper"
 require "./lib/enigma"
+require './lib/shift'
 
 class EnigmaTest <Minitest::Test
   def test_it_exists
@@ -8,16 +9,16 @@ class EnigmaTest <Minitest::Test
     assert_instance_of Enigma, enigma
   end
 
-  def test_create_character_set
-    enigma = Enigma.new
-    expected = ["a", "b", "c", "d", "e", "f", "g", "h", "i", "j", "k", "l", "m", "n", "o", "p", "q", "r", "s", "t", "u", "v", "w", "x", "y", "z", " "]
-    assert_equal expected, enigma.create_character_set
-  end
+  # def test_create_character_set
+  #   enigma = Enigma.new
+  #   expected = ["a", "b", "c", "d", "e", "f", "g", "h", "i", "j", "k", "l", "m", "n", "o", "p", "q", "r", "s", "t", "u", "v", "w", "x", "y", "z", " "]
+  #   assert_equal expected, enigma.create_character_set
+  # end
 
-  def test_create_key
-    enigma = Enigma.new
-    assert_equal 5, enigma.create_key.length
-  end
+  # def test_create_key
+  #   enigma = Enigma.new
+  #   assert_equal 5, enigma.create_key.length
+  # end
 
   def test_create_key_hash
     enigma = Enigma.new
@@ -41,6 +42,7 @@ class EnigmaTest <Minitest::Test
 
   def test_a_shift
     enigma = Enigma.new
+    shift = Shift.new
 
     assert_equal 8, enigma.a_shift("04853", enigma.get_current_date)
   end
